Add files via upload
Classic Snake Game using Python's Turtle Module! 🎮
The nostalgia of the classic Snake game implemented with Python's Turtle module. This project features 3 different classes to manage the snake, food and scoreboard. Perfect for beginners to learn about object-oriented programming and game development in Python.
Check it out, give it a star ⭐, and feel free to contribute!